Update main_ui.py
Created the socket connection in the UI file, added check box for save data into a file. Currently, the code works but the two thread implementation is inneficient and the plot is too slow.
Recode the receivedata and plot into into a same function.
Solve the pause (in the client code)